requirements.txt
streamlit==1.39.0
langchain_community==0.3.5
langchain-huggingface==0.1.2
langchain-groq==0.2.1
pypdf==5.1.0
faiss-cpu==1.9.0
python-dotenv==1.0.1


